control de acceso sin estar logeado como administrador

+++ src/main/java/org/grupo4/practica_integradora_g4/controller/ControladorAdmin.java
@@ -23,36 +23,57 @@ public class ControladorAdmin {
     private CategoriaRepository categoriaRepository;
 
     @GetMapping("/inicio")
-    public String getAdministracion(){
+    public String getAdministracion(HttpSession session, Model model) {
+        if (session.getAttribute("usuario") == null) {
+            return "redirect:/registro/paso1";
+        }
+        model.addAttribute("usuario", session.getAttribute("usuario"));
         return "loginAdmin/administracion";
     }
 
     @GetMapping("/productos")
-    public String listarProductos(Model model) {
+    public String listarProductos(HttpSession session,Model model) {
+        if (session.getAttribute("usuario") == null) {
+            return "redirect:/registro/paso1";
+        }
         model.addAttribute("productos", productoRepository.findAll());
+        model.addAttribute("usuario", session.getAttribute("usuario"));
         return "administrador/productos";
     }
 
     @GetMapping("/productos/add")
-    public String getAgregarProducto(Model model) {
+    public String getAgregarProducto(HttpSession session,Model model) {
+        if (session.getAttribute("usuario") == null) {
+            return "redirect:/registro/paso1";
+        }
         model.addAttribute("nuevoProducto", new Producto());
         model.addAttribute("categorias", categoriaRepository.findAll());
+        model.addAttribute("usuario", session.getAttribute("usuario"));
         return "administrador/addProducto";
     }
 
     @PostMapping("/productos")
-    public String agregarProducto(@Valid @ModelAttribute("nuevoProducto") Producto producto, BindingResult result, Model model) {
+    public String agregarProducto(HttpSession session, @Valid @ModelAttribute("nuevoProducto") Producto producto, BindingResult result, Model model) {
+        if (session.getAttribute("usuario") == null) {
+            return "redirect:/registro/paso1";
+        }
         if (result.hasErrors()) {
             model.addAttribute("categorias", categoriaRepository.findAll());
+            model.addAttribute("usuario", session.getAttribute("usuario"));
             return "administrador/addProducto";
         }
         productoRepository.save(producto);
+        model.addAttribute("usuario", session.getAttribute("usuario"));
         return "redirect:/administrador/productos";
     }
 
     @GetMapping("/productos/eliminar")
-    public String eliminarProducto(@RequestParam String id) {
+    public String eliminarProducto(HttpSession session,Model model, @RequestParam String id) {
+        if (session.getAttribute("usuario") == null) {
+            return "redirect:/registro/paso1";
+        }
         productoRepository.deleteById(id);
+        model.addAttribute("usuario", session.getAttribute("usuario"));
         return "redirect:/administrador/productos";
     }
